+++ src/Api/Billing/Models/Requests/CalculateTaxRequestModel.cs
@@ -4,7 +4,7 @@ public class CalculateTaxRequestModel
 {
     public decimal Amount { get; set; }
 
-    public string BillingAddressPostalCode { get; set; }
+    public string PostalCode { get; set; }
 
-    public string BillingAddressCountry { get; set; }
+    public string Country { get; set; }
 }
